chore: move skip logic to generate
